refactor(sdsapptheme): extract primitive color values into contants/color.ts
to make them consistant with the rest of theme constants
